requirements-dev.txt
-r requirements.txt
pytest-django
pytest-cov
